requirements.txt
django==2.2
djangorestframework==3.9.2
django-storages=1.4.1
